+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Added
 - [Issue 616](https://github.com/aza547/wow-recorder/issues/616) - Add an upload toggle for Retail and Classic recordings.
+- Add a Storage filter toggle, remove the search tags for cloud/disk and the override switch for bypassing cloud videos.
 
 ### Fixed
 - [Issue 642](https://github.com/aza547/wow-recorder/issues/642) - Check daily for updates not just on startup.